+++ packages/manager/apps/dedicated/client/app/dedicated/server/details/server.routing.js
@@ -46,15 +46,10 @@ export default /* @ngInject */ ($stateProvider) => {
       worldPart: /* @ngInject */ (coreConfig) => coreConfig.getRegion(),
 
       breadcrumb: /* @ngInject */ (server) => server.displayName,
+      trackingPrefix: () => 'dedicated::dedicated::server',
 
-      bringYourOwnImage: /* @ngInject */ ($q, $stateParams, Server) =>
-        Server.getBringYourOwnImage($stateParams.productId).catch((error) => {
-          if (error.status === 404) {
-            return null;
-          }
-
-          return $q.reject(error);
-        }),
+      bringYourOwnImage: /* @ngInject */ ($stateParams, Server) =>
+        Server.getBringYourOwnImage($stateParams.productId).catch(() => null),
     },
   });
 };